Committing intermediate progress
