+++ dynd/include/assign.hpp
@@ -26,6 +26,10 @@ extern struct assign_pyarrayscalarobject
   static dynd::nd::callable make();
 } assign_pyarrayscalarobject;
 
+extern struct assign_pyarrayobject : dynd::nd::declfunc<assign_pyarrayobject> {
+  static dynd::nd::callable make();
+} assign_pyarrayobject;
+
 void array_copy_to_numpy(PyArrayObject *dst_arr, const dynd::ndt::type &src_tp,
                          const char *src_arrmeta, const char *src_data);
